AP-756 swagger updates

+++ lib/sib-api-v3-sdk/models/remaining_credit_model_reseller.rb
@@ -31,8 +31,8 @@ module SibApiV3Sdk
     # Attribute type mapping.
     def self.swagger_types
       {
-        :'sms' => :'Integer',
-        :'email' => :'Integer'
+        :'sms' => :'Float',
+        :'email' => :'Float'
       }
     end
